Prevent Error Message when Printing
This enables the printing entitlement when users attempt to print.
Note: it appears that printing is broken; in the print dialog choosing
"Open PDF in Preview" only yields a blank two-page document using
Markoff/README.md as a test. The "Save as PDF" and "Save as Postscript"
commands have no effect either.